requirements.txt
anthropic
boto3
datasets
diskcache
elasticsearch-dsl
hydra-core
ipykernel
matplotlib
omegaconf
openai
polars
pre-commit
python-dotenv
streamlit
typer
wandb
